bug(#7394): read a string literal's escapes somewhere else
`Emissions` was over the God Class line under qulice 0.31.2 — WMC=133
against a threshold of 47, cohesion at zero — so master and every pull
request opened against it were red.

The class was doing three unrelated things, and the one that belongs
least is reading escapes out of a string literal. `\n`, a unicode escape
and the octal `\101` all have to come out as the bytes the program will
hold at runtime, and none of that has anything to do with emitting XMIR.
It moves to `Escapes`, along with the octal byte cap it is the only user
of.

`Emissions.unescapeBytes` went with it rather than staying as a
one-line delegator, so `LnTextBlock` and the string emitter now ask
`Escapes.bytes` directly.

No behaviour changes: 1307 parser tests green, and qulice is clean
across every module.
